requirements.txt
gspread==5.12.4
google-auth==2.35.0
pandas==2.2.3
python-dotenv==1.0.1
openpyxl==3.1.5
